compose-lab/grpc-service/requirements.txt
Flask
grpcio
grpcio-tools
grpcio-reflection
protobuf
pymongo
